libraries/postgres-helpers/pyproject.toml
[tool.poetry]
name = "postgres-helpers"
version = "0.1.1"
description = ""
authors = ["Jakob J. Kolb <jakob.j.kolb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.1"
yoyo-migrations = "^7.3.2"
six = "^1.16.0"
returns = "^0.18.0"
ramda = "^0.7.3"

[tool.poetry.dev-dependencies]
pytest = "^6.0.0"
pytest-xdist = "^2.4.0"
flake8 = "^4.0.1"
black = "^21.10b0"
pyproject-flake8 = "^0.0.1-alpha.2"

[build-system]
requires = ["poetry-core @ git+https://github.com/python-poetry/poetry-core.git@master"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
looponfailroots = 'database'

[tool.flake8]
max-line-length = 120
max-complexity = 10
ignore = "E731, W503"